+++ src/generator/templates/html_users_show.rs
@@ -1,0 +1,8 @@
+pub static TEXT: &'static str = "{% extends \"base_layout\" %}
+
+{% block content %}
+  <h1>User</h1>
+  <div>{{ body.email }}</div>
+  <div>{{ body.id }}</div>
+{% endblock content %}
+";