+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 - Initial release
 - Can generate web application
 - Web application can run on posgresql adapter
+
+# Version 0.0.2 (2018-03-08)
+- Updated README.md
+- Now works with mysql and sqlite adapter as well
+- Users signin and signup works